reddit-bot/requirements.txt
zendriver
openai
python-dotenv
praw
